Began moving units to crate, and the rest to AstroDisplay.
