+++ src/astro_display/mod.rs
@@ -1,0 +1,5 @@
+mod simpel_impl;
+
+pub trait AstroDisplay {
+    fn astro_display(&self) -> String;
+}
