+++ src/units/mass.rs
@@ -7,18 +7,6 @@ use crate::astro_display::AstroDisplay;
 
 use super::DISPLAY_THRESHOLD;
 
-unit! {
-    system: uom::si;
-
-    quantity: uom::si::mass;
-    @kiloton: 1e6; "kt", "kiloton", "kilotons";
-    @megaton: 1e9; "Mt", "megaton", "megatonnes";
-    @gigaton: 1e12; "Gt", "gigaton", "gigatons";
-    @lunar_mass: 7.346e22; "M☽", "lunar mass", "lunar masses";
-    @earth_mass: 5.9722e24; "M🜨", "earth mass", "earth masses";
-    @solar_mass: 1.988416E30; "M☉", "solar mass", "solar masses";
-}
-
 pub enum MassUnit {
     Kilograms,
     Tonne,
